Add dynamic scream chance
A feature from Screambot JS. Screambot becomes more likely to scream in a channel when there is more activity. To do that it needs some state to keep track of a count of recent messages per channel.
This is the second "number of times a certain thing happened recently per channel" thing, so I renamed the first one to make it make more sense in the presence of the other.

Also switched from std::chrono time points to time_t because oh my goodness time points are so verbose
Might change it back now that it's working thought because time_t is only precise to the second and I'd kind of rather it be precise to milliseconds